+++ src/test/java/edu/kis/powp/jobs2d/TestJobs2dPatterns.java
@@ -14,7 +14,7 @@ import edu.kis.powp.jobs2d.events.SelectTestFigureOptionListener;
 import edu.kis.powp.jobs2d.features.DrawerFeature;
 import edu.kis.powp.jobs2d.features.DriverFeature;
 
-public class TestJobs2dSoftPatterns {
+public class TestJobs2dPatterns {
 	private final static Logger logger = Logger.getLogger(Logger.GLOBAL_LOGGER_NAME);
 
 	/**
